Define and use a convience function `fatalf` inside `lwt_wrap_exn`
Since we always used `Log` to log the error, before returning it
as an `Error string, define and use a new function `fatalf` to do
both at once.

Note before we actuallly logged `End_of_file` as "info" but this
was probably a mistake, since it too returned `Error string.

Signed-off-by: David Scott <dave.scott@docker.com>