On Win32, fall back to buffered I/O with `fsync`
Ideally we want to open the file in raw mode to behave more like a
Xen unikernel, where

- data is persisted as soon as `write` returns (and not pending in
  some write back cache)
- data is only copied if we ask for it (and are happy with the
  implications)
- caches are only used if they're linked into the program

On Linux and OSX it's quite straightforward to open files in raw
mode with flags O_DIRECT and F_NOCACHE. On Win32 the APIs are
completely different.

As an initial starting point, this patch adds a Win32 implementation
which opens the file as normal but calls `fsync` after every `write`.
This gains the persistence property, but still causes data to be
copied and cached.

Signed-off-by: David Scott <dave.scott@docker.com>