On OSX use `fcntl(F_FULLFSYNC)` for `flush`
The man page for `fsync` on OS X warns that it's not enough and doesn't
prevent arbitrary re-ordering of writes by the drive. This means that an
application taking care to write structures in a certain order bracketted by
calls to `flush` to maintain invariants will be disappointed by the behaviour
of `fsync`. Switch to `fcntl(F_FULLFSYNC)` as recommended by the manpage.

Signed-off-by: David Scott <dave@recoil.org>