Add another configurable flush behaviour
Previously we had 2 options on macOS:

- do nothing on flush (fast and unsafe)
- call fcntl(F_FULLFSYNC) to flush data all the way to the drive (slow
  but reliable)

This patch adds an additional option to flush using `fsync` which will
flush to the OS, but possibly not all the way to the drive. This is much
faster than `fcntl` but means that if the host OS crashes we might
corrupt data. However if the host OS crashes then we might corrupt data anyway.

Signed-off-by: David Scott <dave@recoil.org>