On read/write failure, mark the cached seek_offset as undefined
Previously we could complete part of a read or write, advancing the
real file offset, but fail the overall operation so the t.seek_offset
remains at the original value.

This patch catches all failures from read/write/readv/writev and sets
t.seek_offset to -1L which means the next call to `seek_already_locked`
will conclude that a real call to `lseek` is needed.

Signed-off-by: David Scott <dave@recoil.org>