Detect the sector size of block devices
Previously we always hardcoded the value of 512, which would cause
problems if using unbuffered I/O (the default) on modern 4k disks.

This patch adds a binding to the `ioctl` which queries the sector
size of block devices. Unfortunately there is no obvious way to discover
the "sector size" so we still hardcode 512 in that case. Perhaps we
should switch to regular (buffered) file I/O by default to avoid this
issue?

Signed-off-by: David Scott <dave@recoil.org>