Use regular buffered I/O by default
Previously we would use unbuffered I/O which insists that all writes
are persisted to the disk before returning success-- this is a good
default for databases where data loss is completely unacceptable.

In practice most applications only require write to hit the OS caches,
and will tolerate data loss if the host OS occasionally crashes-- this
is the default of the `write` API as used by most applications.

This patch switches to buffered I/O by default i.e. we now have the
same default as the underlying Unix. This means that

- I/O will not fail if buffers are misaligned (useful given the
  alignment constraints are hard to determine in practice)
- I/O will be much faster

Signed-off-by: David Scott <dave@recoil.org>