Fix TRIM/DISCARD of small regions
We present a virtual sector size of 512 to the VM but are likely using
a 4096 byte native sector size on the local SSD, therefore we have to
take care to align calls to `fcntl(F_PUNCHHOLE)`.

Previously we would align up the `fp_offset` to the native sector size
and then compute the length of the unaligned initial region we need
to manually zero by `aligned_offset - fp_offset`. Unfortunately this
fails to take into account that the `fp_length` might be smaller than
this, for example:

assume `fp_offset = 1024` and `fp_length = 512` (i.e. TRIM one sector)
=> `aligned_offset = 4096`
=> `len_to_zero = aligned_offset - fp_offset = 3072`

and we will erase `3072` bytes even though the request was for `512`.
We will then go on to calculate a -ve number for the `fcntl(F_PUNCHHOLE)`
call which will return `EINVAL`, but damage has already been done.

This patch caps the size of the initial region to `fp_length`.

Signed-off-by: David Scott <dave@recoil.org>